feat: `Lotto` 클래스에 로또 번호의 검증 책임 부여
