+++ src/model/Lotto.js
@@ -1,0 +1,24 @@
+import Validator from '../validators/Validator.js';
+
+class Lotto {
+  #numbers;
+
+  constructor(numbers) {
+    this.#validate(numbers);
+    this.#numbers = numbers;
+  }
+
+  #validate(numbers) {
+    Validator.validateLottoNumbers(numbers);
+  }
+
+  getWinningNumbers() {
+    return this.#numbers;
+  }
+
+  getLotto() {
+    return this.#numbers;
+  }
+}
+
+export default Lotto;
